Use long as fk in dataset version

+++ src/main/java/marquez/api/DatasetVersion.java
@@ -14,13 +14,13 @@ import javax.validation.constraints.NotNull;
 })
 public abstract class DatasetVersion {
   @NotNull private final Timestamp createdAt;
-  @NotNull private final Integer datasetId;
+  @NotNull private final long datasetId;
   @NotNull private final String schemaUri;
 
   @JsonCreator
   public DatasetVersion(
       @JsonProperty("created_at") final Timestamp createdAt,
-      @JsonProperty("dataset_id") final Integer datasetId,
+      @JsonProperty("dataset_id") final long datasetId,
       @JsonProperty("schema_uri") final String schemaUri) {
     this.createdAt = createdAt;
     this.datasetId = datasetId;
@@ -31,7 +31,7 @@ public abstract class DatasetVersion {
     return createdAt;
   }
 
-  public Integer getDatasetId() {
+  public long getDatasetId() {
     return datasetId;
   }
